002_NLP/requirements.txt
protobuf==6.32.0
python-dotenv==1.1.1
pyttsx3==2.99
